# back/temp_swagger.json
{"openapi":"3.0.0","info":{"title":"OIEM Abastible Backend API","version":"1.0.0","description":"API Documentation for OIEM Abastible System (IEEE Trace: REQ-013)","contact":{"name":"Abastible AI Agent"}},"servers":[{"url":"http://localhost:4000/api","description":"Development server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"User":{"type":"object","required":["name","email","role"],"properties":{"id":{"type":"integer","description":"The auto-generated id of the user"},"name":{"type":"string","description":"The name of the user"},"email":{"type":"string","description":"The email of the user"},"role":{"type":"string","enum":["admin","administrador_contrato","contratista_admin","contratista_user"],"description":"The role of the user"},"activo":{"type":"boolean","description":"Whether the user is active"}},"example":{"id":1,"name":"John Doe","email":"john@example.com","role":"admin","activo":true}},"LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}},"example":{"email":"admin@abastible.cl","password":"User123*"}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string","description":"JWT Access Token"},"user":{"$ref":"#/components/schemas/User"}}},"Programa":{"type":"object","properties":{"id":{"type":"integer"},"nombre":{"type":"string"},"descripcion":{"type":"string"},"activo":{"type":"boolean"}},"example":{"id":1,"nombre":"OIM Distribución Granel","descripcion":"Programa de Operación Integrada","activo":true}},"Elemento":{"type":"object","properties":{"id":{"type":"integer"},"numero":{"type":"string"},"nombre":{"type":"string"},"orden":{"type":"integer"}},"example":{"id":1,"numero":"1","nombre":"Liderazgo","orden":1}},"Actividad":{"type":"object","properties":{"id":{"type":"integer"},"codigo":{"type":"string"},"actividad":{"type":"string"},"descripcion":{"type":"string"},"frecuencia":{"type":"string","enum":["mensual","trimestral","semestral","anual","cuando_aplique"]},"requiere_evidencia":{"type":"boolean"}},"example":{"id":1,"codigo":"1.1","actividad":"Reuniones gerenciales","descripcion":"Realizar reunión mensual","frecuencia":"mensual","requiere_evidencia":true}},"Registro":{"type":"object","properties":{"id":{"type":"integer"},"periodo":{"type":"string","format":"date"},"porcentaje_cumplimiento":{"type":"number","format":"float"},"estado_auditoria":{"type":"string","enum":["pendiente","auditando","auditada_terreno","auditada_sistema","reabierto"]},"auditado":{"type":"integer","description":"0 or 1"},"cerrado":{"type":"integer","description":"0 or 1"}},"example":{"id":10,"periodo":"2026-01-01","porcentaje_cumplimiento":85.5,"estado_auditoria":"pendiente","auditado":0,"cerrado":0}},"Error":{"type":"object","properties":{"message":{"type":"string"},"error":{"type":"string"}},"example":{"message":"Recurso no encontrado","error":"Not Found"}}}},"security":[{"bearerAuth":[]}],"paths":{"/programas":{"get":{"summary":"List all programs","tags":["Programas"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of programs","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Programa"}}}}}}},"post":{"summary":"Create a new program","tags":["Programas"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Programa"}}}},"responses":{"201":{"description":"Program created"}}}},"/elementos":{"get":{"summary":"List all elements","tags":["Elementos"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of elements"}}},"post":{"summary":"Create element","tags":["Elementos"],"security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Elemento"}}}},"responses":{"201":{"description":"Created"}}}},"/actividades":{"get":{"summary":"List activities","tags":["Actividades"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of activities","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Actividad"}}}}}}},"post":{"summary":"Create activity","tags":["Actividades"],"security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Actividad"}}}},"responses":{"201":{"description":"Created"}}}},"/registros":{"get":{"summary":"List monthly registers (filtered by role)","tags":["Registros"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of registers","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Registro"}}}}}}},"post":{"summary":"Create a new register","tags":["Registros"],"security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"periodo":{"type":"string","format":"date"},"contratista_asignacion_id":{"type":"integer"}}}}}},"responses":{"201":{"description":"Register created"}}}},"/registros/{id}/auditar":{"post":{"summary":"Start auditoria for a register","tags":["Registros"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Auditoria started"}}}},"/evidencias":{"post":{"summary":"Upload evidence file","tags":["Evidencias"],"security":[{"bearerAuth":[]}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"archivo":{"type":"string","format":"binary"},"registro_actividad_id":{"type":"integer"}}}}}},"responses":{"201":{"description":"File uploaded"}}},"get":{"summary":"List evidences","tags":["Evidencias"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of evidences"}}}},"/hallazgos":{"get":{"summary":"List findings","tags":["Hallazgos"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of findings"}}},"post":{"summary":"Create finding","tags":["Hallazgos"],"security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Finding created"}}}},"/compromisos":{"get":{"summary":"List commitments","tags":["Compromisos"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of commitments"}}}},"/reaperturas":{"get":{"summary":"List reopen requests","tags":["Reaperturas"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of requests"}}},"post":{"summary":"Create reopen request","tags":["Reaperturas"],"security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Request created"}}}},"/dashboard/kpis":{"get":{"summary":"Get main KPIs","tags":["Dashboard"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Dashboard KPIs"}}}},"/dashboard/cumplimiento":{"get":{"summary":"Get compliance stats","tags":["Dashboard"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Compliance data"}}}},"/reportes/registro/{id}/pdf":{"get":{"summary":"Generate PDF report for register","tags":["Reportes"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"PDF file","content":{"application/pdf":{"schema":{"type":"string","format":"binary"}}}}}}},"/usuarios":{"get":{"summary":"List all users","tags":["Usuarios"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"summary":"Create a new user","tags":["Usuarios"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"User created"}}}}},"tags":[]}